Add files via upload
The main aim of this study is twofold: firstly, to conduct an examination of the labor market in the United Kingdom in the year 2004, and secondly, to investigate the interrelated dynamics between individual job satisfaction and group cohesiveness. The objective of this research is to examine the factors that contribute to elevated levels of morale and collaboration within an organization's workforce. To achieve this, the study utilizes the extensive dataset offered by the 2004 Workplace Employment Relations Survey (WERS).